Many improvements and fixes.
Main fixes are that of fixing settings menu, adding service class to run in the background (should stop the app not working after a few hours) and a lot of small fixes relating to those.

Some functionality currently broken due to the change of the app structure, but the overall app now works a lot better.